fixed mindist aabb

+++ lbvh/aabb.cuh
@@ -62,6 +62,7 @@ inline float mindist(const aabb<float>& lhs, const float4& rhs) noexcept
     const float dz = ::fminf(lhs.upper.z, ::fmaxf(lhs.lower.z, rhs.z)) - rhs.z;
     return dx * dx + dy * dy + dz * dz;
 }
+
 __device__ __host__
 inline double mindist(const aabb<double>& lhs, const double4& rhs) noexcept
 {
@@ -80,18 +81,20 @@ inline float minmaxdist(const aabb<float>& lhs, const float4& rhs) noexcept
     float3 rM_sq = make_float3((lhs.upper.x - rhs.x) * (lhs.upper.x - rhs.x),
                                (lhs.upper.y - rhs.y) * (lhs.upper.y - rhs.y),
                                (lhs.upper.z - rhs.z) * (lhs.upper.z - rhs.z));
-    if((lhs.upper.x + lhs.lower.x) * 0.5 < rhs.x)
+    
+    if((lhs.upper.x + lhs.lower.x) * 0.5f < rhs.x)
     {
         thrust::swap(rm_sq.x, rM_sq.x);
     }
-    if((lhs.upper.y + lhs.lower.y) * 0.5 < rhs.y)
+    if((lhs.upper.y + lhs.lower.y) * 0.5f < rhs.y)
     {
         thrust::swap(rm_sq.y, rM_sq.y);
     }
-    if((lhs.upper.z + lhs.lower.z) * 0.5 < rhs.z)
+    if((lhs.upper.z + lhs.lower.z) * 0.5f < rhs.z)
     {
         thrust::swap(rm_sq.z, rM_sq.z);
     }
+    
     const float dx = rm_sq.x + rM_sq.y + rM_sq.z;
     const float dy = rM_sq.x + rm_sq.y + rM_sq.z;
     const float dz = rM_sq.x + rM_sq.y + rm_sq.z;
@@ -107,6 +110,7 @@ inline double minmaxdist(const aabb<double>& lhs, const double4& rhs) noexcept
     double3 rM_sq = make_double3((lhs.upper.x - rhs.x) * (lhs.upper.x - rhs.x),
                                  (lhs.upper.y - rhs.y) * (lhs.upper.y - rhs.y),
                                  (lhs.upper.z - rhs.z) * (lhs.upper.z - rhs.z));
+
     if((lhs.upper.x + lhs.lower.x) * 0.5 < rhs.x)
     {
         thrust::swap(rm_sq.x, rM_sq.x);
@@ -119,10 +123,11 @@ inline double minmaxdist(const aabb<double>& lhs, const double4& rhs) noexcept
     {
         thrust::swap(rm_sq.z, rM_sq.z);
     }
+
     const double dx = rm_sq.x + rM_sq.y + rM_sq.z;
     const double dy = rM_sq.x + rm_sq.y + rM_sq.z;
     const double dz = rM_sq.x + rM_sq.y + rm_sq.z;
-    return ::fminf(dx, ::fminf(dy, dz));
+    return ::fmin(dx, ::fmin(dy, dz));
 }
 
 template<typename T>
